+++ webapp/src/components/video_fallback.tsx
@@ -1,7 +1,6 @@
-import React, { useEffect, useState } from 'react';
+import React, {useEffect, useState} from 'react';
 
-import { getVideoUrl } from '../utils';
-import { Client } from '../client/client';
+import {Client} from '../client/client';
 
 import LinkFallback from './link_fallback';
 
@@ -13,7 +12,7 @@ interface Props {
 
 const VideoFallback = ({fallback, videoProps, children}: Props) => {
     const [text, setText] = useState<string>('')
-    
+
     const videoKey = videoProps.video_key as string;
     const start = videoProps.start as number;
     const length = videoProps.length as number;
@@ -33,7 +32,7 @@ const VideoFallback = ({fallback, videoProps, children}: Props) => {
             getNodeData(nodeID)
         }
     }, [fallback])
-    
+
     return (<LinkFallback fallback={fallback} link={link} text={text}>
         {children}
     </LinkFallback>)
@@ -41,3 +40,17 @@ const VideoFallback = ({fallback, videoProps, children}: Props) => {
 
 
 export default VideoFallback;
+
+const getVideoUrl = (videoKey: string, start?: number, length?: number) => {
+    const urlSearchParams = new URLSearchParams();
+    urlSearchParams.append('v', videoKey);
+    if(start){
+        urlSearchParams.append('start', start.toString());
+    }
+    if(length){
+        const end = ((start || 0)+length)
+        urlSearchParams.append('end', end.toString());
+    }
+    return `https://www.youtube.com/watch?${urlSearchParams.toString()}`
+};
+
